basic functionality
The python script starts the measurement after the arduino gives a go
command. The arduino prints out raw sensor data from the MAX31865
sensor.  The python program interprets the incoming data and translates
it to a temperature. The data is plotted using matplotlib and is stored
to a csv using pandas.